Extract out LogSeverity into own class
